leffa_test/requirements.txt
gradio
gradio-client
Pillow
opencv-python 